refactor(hmr): move the business platform out of the minimal core
feat/hot-update-mvp should carry mechanism only: drop terminal.ts, the
platform dev-push entry, and the terminals business (both platform.ts's
`terminals` child and the /terminals* routes) from the packaged default.
The packaged platform is now a bare stub (motd context + info()); a real
business platform arrives over HTTP and, going forward, is dispatched
generically rather than wired route-by-route.

watch-push.mjs tolerates a missing platform entry (frontend-only push
cycle) so the dev loop still works on a mechanism-only checkout. The
terminal-swap proof in hmr.test.ts moves to hmr-business-platform.test.ts,
parked with describe.skip until the business platform is restored.
